BREAKING CHANGE: bufferline flow and mappings
change bufferline plugin for lualine, and now
the buffer list and nvim information is in the same
location.

Additionally, it was deleted unnecessary code from utils.lua
and now we are using the nvim api for the keybinds directly.
Also, we added which key and some descriptions to the keybinds